requirements.txt

snowflake-snowpark-python
streamlit
pandas
